+++ openrouteservice/src/main/java/heigit/ors/routing/graphhopper/extensions/reader/osmfeatureprocessors/OSMWayFilter.java
@@ -1,0 +1,8 @@
+package heigit.ors.routing.graphhopper.extensions.reader.osmfeatureprocessors;
+
+import com.graphhopper.reader.ReaderElement;
+import com.graphhopper.reader.ReaderWay;
+
+public abstract class OSMWayFilter implements OSMFeatureFilter {
+
+}